new version

+++ modules/remove/story.bash
@@ -4,11 +4,24 @@ debug=$(config debug)
 
 volume_group=$(config vg)
 logical_volume=$(config lv)
-use_force=$(config force)
+use_force=$(story_var force)
 
-if [[ $use_force == yes ]]; then
-  remove_opts+=" -f"
+remove_opts+="-y "
+
+if [[ $use_force == true ]]; then
+  remove_opts+="-f "
+  umount "/dev/mapper/$volume_group-$logical_volume" 2>/dev/null
 fi
 
-lvremove $remove_opts $volume_group/$logical_volume 
-vgremove $remove_opts $volume_group 
+check_lvs
+
+if [[ `echo $lvs | grep -w $logical_volume` ]]; then
+  lvremove $remove_opts $volume_group/$logical_volume 
+fi
+
+check_lvs
+
+if [[ ! $lvs ]] ; then
+  echo "No logical volume left. Try to remove a volume group."
+  vgremove $remove_opts $volume_group 
+fi
